Fuji compiles under Emscripten!
Clang support.
Lots of bug fixes thanks to Clang!
Removed non-VBO paths from OpenGL renderer.
Added 'Web' platform and driver.
Renamed platform MF_PC -> MF_Windows.
Removed the bloody new/delete operators.

+++ Fuji/Source/Asset/MFIntStringTable.cpp
@@ -328,9 +328,9 @@ MF_API MFIntStringTable *MFIntStringTable_CreateFromFileInMemory(const void *pMe
 		table.language = language;
 
 		// set some useful details
-		if(enumList == a)
+		if(enumList == (int)a)
 			pStringTable->enumList = l;
-		if(englishList == a)
+		if(englishList == (int)a)
 			pStringTable->englishList = l;
 		++l;
 	}
